refact: add videoutil.py
Move video controller to videoutil.py
